Terminology/syntax adjustment.
All references to metabolic/non-metabolic proteins have been changed to 
functional/non-functional proteins. Non-functional proteins are proteins 
that either have height or width of zero, or are missing codons for one 
of the three necessary features (W, H, M). Referring to them as 
metabolic creates potential confusion with fitness components, where now 
we have metabolic component of the fitness as well as non-metabolic 
(such as secretion) components. In the future, metabolic gene will refer 
to a gene that codes for a protein on the metabolic segment of the axis. 

The changes included variables that contain _met_ , now _fun_, as well as 
_nm_ (number of metaboilc), now _nf_.