Good afternoon,
I have made two changes about the plasmid transfer. First, in the ae_individual.cpp, I have change the inject_GU function to make it add the GU at the end of the list of the GU. So when the program is getting rid of the outnumbered Gu(s) in the cell, it is keeping the last arrived (and not the first arrived as it was the case). So now, a plasmid can really transfer itself from cell to cell more than one time per generation, thanks to the parameter NB_HORIZONTAL_TRANS.

The second change is the replacement of 'evaluate' by 'reevaluate' at the line 846 of ae_population.cpp . Indeed, previously the individual was never reevaluated (because it has already evaluated before), and nothing was never changed. The previous version was doing that: a plasmid's selection was function of the characteristics of the previous plasmid of the cell; not of the characteristics of the plasmid which replaced it.

Thank you,

Renaud Dessalles