1. Add a simple test system.
For a set of param files, we:

    * run aevol_create
    * run aevol_run for 10 generations
    * check that generation 10 has been saved in populations
    * check that stats/stat_fitness_glob contains the right number of lines

For now we run this test for a basic run and for a run with spatial structure.

To add a param file to test:

    * Put the param file in the directory "tests/", with a name describing what it does, **ending with ".in"** 
    * Modify Makefile.am: on the line "TESTS = ", add your param file.
    * Run autoreconf or equivalent so your modification to Makefile.am is taken into account 

It would be helpful if everybody could add a typical param file with the features he uses in "tests/" directory.

To run the tests, just type 'make check'. If you configured 'with--cppunit', then both unitary tests and these validation tests will be run.

2. Fix build system in doc folders: remove Makefile and add Makefile.in to svn.