============================== trunk ============================= BUGFIX (resolves 281 and 282)
Bug in gene position

This bug happenned whenever a promoter at the end of the genome on the LEADING strand transcribes a gene that is at the
beginning of the genome
(or a promoter at the beginning on the LAGGING strand transcribes a gene that is at the end of the genome)

The position of the protein is then out of the genome bound (> genome len when LEADING or < 0 when LAGGING)
If this gene was also transcribed on another rna that didn't satisfy the above constraint, it wasn't recognized as the same gene.


Either bugs had no effect whatsoever on the fitness and hence on the outcome of evolution.
It could however lead to erroneous statistics regarding the number of RNAs a gene is transcribed onto.