================= IN TRUNK : (RATHER MINOR) BUGFIXES ================
This revision fixes the bugs listed as #283, #284, #285, #285
on the gforge website. See details below.


Bug #283:
--------

The test system introduced in r369 to check that aevol does not crashes in various conditions (basic.simple, space.simple, secretion.simple) did not work under Mac OS X due to a non-POSIX option of the command find in the script simple.sh. I fixed it by replacing -executable by the POSIX-compliant option -perm -u+x. Note that 'make check' now works on my MacOS (where my automake version is 1.13) but still does not work on my Linux Ubuntu (where automake version is 1.11) ==> Antoine, if you understand why...


Bug #284: undefined behavior when there is no terminator in the genome
--------

After a big deletion, it can happen that a genetic unit does not contain any teminator anymore. There can however be a promoter somewhere. The behavior of the program was different depending on the strand where the promoter was. If it was on the lagging strand, the RNA was supposed to be as long as the whole genetic unit, and could carry coding sequences. If the promoter was on the leading strand, the length of the RNA was left as is, that is, either to -1 for generation 0, or to the length  inherited from the parent, which made no sense anymore once the terminator has disappeared.

After talking with Guillaume, we decided that it is best that no RNA is produced in this case (no terminator = incomplete gene, assumed to be non-functional), and a fortiori no coding sequence should be searched for there. However, the promoter (actually, an ae_rna object) should not be deleted, because we still need to transmit it to the offspring, in case a terminator is recreated by a mutation. Thus the ae_rna object should be kept in memory but with a transcript_length equals to -1. This will prevent the search for coding sequences in this ae_rna.

Hence in ae_genetic_unit::do_transcription, there is now a call to rna->set_transcript_length(-1) if no terminator was found at all on the genetic unit.



Bug #285: "Barrier" events logged twice
---------

When a duplication or a deletion would cause the genetic unit to be smaller than the min size or larger than the max size, the mutation is not performed and (if the BARRIER option was chosen in the logs), an entry should be written in log_barrier.out. Some deletion events were actually logged twice, once in ae_dna::do_deletion(void) and once in ae_dna::do_deletion(int, int, int).The log entry in do_deletion(void) was correct. The other entry should not have been issued, because it pertains to events that left the genetic unit in an acceptable state (_length > minlength). It was issued for deletions that left the genetic with a length between 10bp (default minlength) and the promoter length (22bp). Such a deletion is accepted and performed but a log entry was issued.

This is now corrected in the various ae_dna methods.


Bug #286 :  Log files incorrectly regenerated when resuming a run
---------

When resuming a run, log files are regenerated by copying the header of the former log file and its entries until the generation chosen to resume the run. During this copy, (1) the first entry was skipped, and (2) the entries for the resuming generation were copied, and thus ended up duplicated.

This is now corrected in ae_logs.cpp.