========================= IN TRUNK =======================
----------------
!!! WARNING !!!
----------------

In this revision, the file format of the mutations is modified 
(see details below). If you update your local copy to this
revision, you will not be able to read your tree files
anymore. Before running the "svn up" command, please make sure 
you have a backup of the code you used to produce your 
simulation data. 



Details: 
--------

This is the proper fix for Bug #300: "Missing length 
data in the tree files for the rearrangements". 
The length of the rearrangements is now written on disk 
in the tree files. Now an ae_mutation object contains
all the information necessary to describe the mutation, 
regardless of how the object was constructed (either freshly 
created by a call to "report_...", or loaded from a file).