======================== IN TRUNK ======================
NEW:
----

Updated the fixed_mutations post-treatment to have the number 
of genes affected by each mutation in the output file. More
precisely, three new columns are produced:
- the number of coding RNAs possibly disrupted by a switch, a small indel
  or a breakpoint
- the number of coding RNAs entirely included in a rearranged segment
- in the case of a transfer by replacement, the number of genes that 
  were entirely included in the replaced segment.

This post-treatment does not work if translocations were allowed
between different genetic units, hence it is not usable in practice if
plasmids are allowed.

I have tested it in the basic case, with and without environmental
variation. I have also tested it in the case of transfer with replacement
"with close points" (Berenice's case). 


BUGFIX:
------

I also fixed a bug in the same post-treatment: the fitness impact
of some mutations was wrong if there was some environmental variation, 
because the individual was not immediately reevaluated after the
environmental variation.

==========================================================