================== IN TRUNK ===================
BUGFIX
------

This commit fixes bug #304: "create_eps produces a seg fault".
Due to changes from aevol-3 to aevol4, we attempted to delete 
the environment twice at the end of the program. 
This post-treatment is now fully functional again. I tested it 
with and without spatial structure.

I have also cleaned the output and removed obsolete comments 
in this post-treatment.