========================== IN TRUNK =======================
NEW:The gene_families post-treatment issues the detailed history of each gene family on the lineage of a given individual (providing its lineage file, see section). A gene family is defined here as a set of coding sequences that arised by duplications of a single original gene. The original gene, called the root of the family, can either be one of the genes in the initial ancestor, or a new gene created from scratch (for example by a local mutation that transformed a non-coding RNA into a coding RNA). The history of gene duplications, gene losses and gene mutations in each gene family is represented by a binary tree. The program starts by loading the initial genome at the beginning of the lineage and by tagging each gene in this initial genome. Each of these initial genes is marked as the root of a gene family. Then, each mutation recorded in the lineage file is replayed and the fate of all tagged genes is followed and recorded in their respective families. When a gene is duplicated, the corresponding node in one of the gene trees becomes an internal node, and two children nodes are added to it, representing the two gene copies. When a gene sequence is modified, the mutation is recorded in its corresponding node in one of the gene trees. When a gene is lost, the corresponding node in one of the gene trees is labelled as lost. When a new gene appears from scratch, i.e. not by gene duplication, it becomes the root of a new gene tree. Environmental variations are also replayed exactly as they occured during the main run. 

When all mutations have been replayed, several output files are written in a directory called gene_trees. Two general text files are produced. The file called gene_tree_statistics.txt contains general data on each gene family, like its creation date, its extinction date, or how many nodes it contained. The file called nodeattr_tabular.txt contains information about each node of each gene tree, like when it was duplicated or lost or how many mutations occurred on its branch. In addition, for each gene tree, two text files are generated: a file called genetree******-topology.tre contains the topology of the gene tree in the Newick format, and a file called genetree******-nodeattr.txt that contains the detailed list of events that happened to each node in the tree file, before it was either duplicated or lost. 

Usage: ae_misc_gene_families [-c | -n] [-t tolerance] -f lineage_file