========================= IN TRUNK ==========================
BUGFIX: Some gene tree nodes were missing in the output of
the gene_families post-treatment. This happened when a
duplicated segment inserted itself in the original copy.
In this case, the duplication is reported in the gene tree,
but the left child is immediately lost. Hence, some internal
nodes can have a single child instead of two, and this was 
not properly handled by the output methods.