======================== IN TRUNK ========================
BUGFIX: The gene_families post-treatment could produce
an error ("Error: the protein [memadress] should not be found 
twice in the tree") in some rare cases, when (i) the start 
position of a gene was deleted and (ii) there was another
gene starting exactly n bases further, where n is the 
size of the deletion. This happened because the gene was 
considered modified rather than lost when its start was deleted,
if the stop was not deleted. This is now corrected, a gene
is now considered lost when a deletion encompasses its start,
regardless of whether the deletion also encompasses the stop 
or not.


BUGFIX: Also in gene_families, the genes tree statistics 
(number of leaves, number of internal nodes, etc) were wrong
in the rare case where a gene is duplicated and immediately
killed by the insertion of the duplicate (when the duplicate 
inserts itself in the original copy, thereby killing it).


Added more debug info in ae_individual::print_protein_list.

Cosmetic change in the comments of ancstats.