Working on "detached mode"
Many things are intricated in aevol. and sometimes way too much.
For example many things rely on the existance of an experiment manager. However, this assumption is getting problematic since we are now often fooling around with experiments, e.g. mixing them together.

It is now possible to manipulate a population that is "detached" i.e. not attached to a given exp manager.
It is very likely that this doesn't work in some cases but it does work in at least some cases (for example I have implemented a post-treatment -- or intermediate treatment ;-p -- that mixes 2 populations together. This works fine).