===================  IN TRUNK =================
This revision fixes Bug #309: "aevol post treatments 
produce a seg fault when remotely run through ssh"

Details: When aevol was compiled with X11 on a 
machine, and if we access that machine through ssh 
to run post-treatments like lineage or ancstats or
create_eps or robustness (etc), we get a segmentation 
fault because for some reason they try to access an 
X sever. This is because they try to create a new 
ae_exp_manager_X11, which is weird because they 
actually do not need the X11 functionalities at all.

This porblem is now fixed. The post treatments that
do not need X11 windows just create an ae_exp_manager,
not a ae_exp_manager_X11.