==================== IN MASTER BRANCH =====================
BUGFIX: Fixed a memory leak in the destructor of
ae_output_manager. The _stats, _tree and _dump
attributes were not destroyed even when they were
not NULL.

This was a rather minor bug since only one instance
of ae_output_manager exists during a simulation.
This bugfix just ensures a cleaner exit of the
program, it should not affect the simulation results.