[MASTER] NEW: Mutagenesis post-treatment tool
This "new" mutagenesis tool is actually a resurrected version of a
program David wrote for his PhD and which originally simulated
similarity-based rearrangements only. I have ported it to version 4
and, more importantly, I have generalized it to all types of
mutations, local or rearrangements, with random or similarity-based
breakpoints.

Lateral gene transfer, however, is not available yet for this tool.
More precisely, this tool can be used to analyze a population evolved
with lateral gene transfer, but only local mutations or
intrachromosomal rearrangements can be performed in the mutagenesis
analysis (for now).

The mutagenesis can be performed for populations evolved on a
spatial grid, with or without secretion. If secretion was enabled
during the evolutionary run, the mutagenesis tool will report not
only the metabolic error of the mutants, but also their secretion
error.

By contrast, the tool is not ready yet for plasmids, because
right now trabnslocations between different genetic units is
not managed yet. I will work on that in the following days.

I added a man page for this tool, as well as a dedicated section
in the user guide. I report below the main information necessary
to use the tool.

Usage:
aevol_misc_mutagenesis -g GENER [-i INDEX | -r RANK]
                       [-m MUTATIONTYPE] [-n NBMUTANTS]

This tool creates and evaluates single mutants of an individual
saved in a backup,  by default the best of its generation. Use
option -g to specify the generation number contanining the individual
of interest. There must have been a backup of the population at this
generation.

Use either the -r RANK or the -i INDEX option to select another
individual than the best one.

The type of mutations to perform must be specified with the -m option.
Choose 0 to create mutants with a point mutation, 1 for a small
insertion, 2 for a small deletion, 3 for a duplication, 4 for a large
deletion, 5 for a translocation or 6 for an inversion.

For the point mutations, all single mutants will be created and
evaluated. For the other mutation types, an exhaustive mutagenesis
would take too much time, hence only a sample of mutants (1000 by
default) will be generated. Use option -n to specify another sample
size.

The output file will be placed in a subdirectory called
analysis-generationGENER.