Faster initialization of plasmid sequences with good genes
Sometimes we want initialize a run with a chromosome sequence and a plasmid sequence both containing at least one beneficial gene. To do so, in param.in, we set INIT_METHOD to ONE_GOOD_GENE (forcing the chromosome to have at least one good gene) and PLASMID_INITIAL_GENE to 1 (forcing the plasmid to also carry a good gene, but not a copy from the chromosomal one). This combination of options was so slow that aevol_create seemed to loop forever. That happened because we tried to find a random individual with, by chance, two genes, one on the chromosome and one on the plasmid. The probability for this to happen was very very low. Instead, I changed the algorithm to make first find a good chromosome, and then (keeping that particular chromosome) a good plasmid.