Fixed a bug regarding the minimal length of genetic units
Aevol runs with plasmids could end up stuck in a infinite loop due to genetic units reaching a length of 0 bp. That happened in particular for plasmids, if the parameter PLASMID_MINIMAL_LENGTH was omitted in the param.in. Comments in the code said in this case, this minimal length is set to the minimal total genome size (10 bp by default), but this was actually not implemented.