fix slow X on OSX since Yosemite update
The OSX Yosemite (10.10) update came with a dramatic
slowdown in aevol X displays. The problem comes from
slow calls to Xlib's XAllocColor() in
ae_X11_window::get_pixel(...).

Since debugging the problem would require to trace
the program down to the Xlib, it has been solved
with a workaround by memoizing the calls to get_pixel()
so that XAllocColor() be actually called only when
requiring a new color.

This solutions costs a bit or memory (around 20Kb on
the basic example) but the display's colors are
preserved and the speed is back to decent.