wrap ae_* in namespace aevol
The name convention ae_* for aevol-related classes and functions comes
from C language lacking namespaces. It was meant to prevent name
clashes. Since C++ provides namespaces, this commit wraps everything
inside an "aevol" namespace.

This means that there is no need anylonger to name classes according to
ae_*. But any new class needs to be within the aevol namespace if it is
part of aevol or to start with a "using" clause if it relies on aevol.