Use correct (param_loader) PRNG to initialize sequences
The new way to initialize the sequences of genetic units (to allow
for a faster initialization of the plasmid sequence) yielded
different sequences when compared to the old way, even in the
absence of plasmid. This is because in my previous commits, I used
the _mut_prng of ae_individual instead of the _prng of param_loader
to draw the random sequences. This is now fixed.