fix overoptimization genome < promoter + terminator
The code used to be optimized to bypass phenotype computation when the
genome was smaller than the length of a promoter plus the length of a
terminator. But terminators can overlap promoters when they share a
chunk of DNA. Viruses showed that this actually happens. The condition
has thus been relaxed to bypasse genomes that are smaller than a
promoter.