add method to check if environment attributes are set
There were several calls to ae_environment getters (get_gaussians and
get_custom_points) only to check if the attributes were set. Added to
predicate accessors to check just this.

Well, the true story behind this is that Eclipse got me to believe
that *all* the calls were done to check which attribute was set. So it
would have been a greater win. However, this change will be useful to
prevent passing lists around when they are std::lists.