All the PRNGs are now handled through smart pointers
PRNGs are a typical example of entities whose ownership depends
on the parameters provided to the system. There can either be a
PRNG per individual (or grid-cell) or a single one may be used
for the whole population.
To avoid having to hand-manage these cases (and those I still
haven't thought of), they are now handled through std::shared_ptr

For consistency the other PRNGs are now handled through
std::unique_ptr