fix duplicated ae_rna in _rna_list
A bug that drove me crazy. `_rna_list[LEADING]` was referencing the same
ae_rna cell twice. First part of the bug was that I had forgotten to
remove the extracted ae_rna cells from the list. Second part was that I
was removing them from a *copy* of the list.