Add empirical test for transcription and translation
With the current data, this test checks that all the transcription
and translation process is consistent with aevol-4.4 on the wild types
generated by Vincent for Santi.

WARNING: this means these tests are run only on short genomes (for now). This
will need to be done for longer genomes as well.

WARNING2: I have commented the tests on m, w and h values because, even
though they can be done (using EXPECT_NEAR with an adequate precision
value), they do not pass (the rest does, hooray !). This means there is a
bug somewhere in the folding process.