Refactor tree management
The responsability of the creation of the tree was shared between
different entities (the tree of course, but also the individuals
and their dna).
This made it really difficult to deal with.

The solution presented here isn't optimal (yet ?). In particular,
trees had to be temporarily disabled for experiments with plasmids.
The tree now handles itself more independently. It is filled with
naive ReplicationReports which will have to be :
  - initialized at the beginning of a replication
  - filled in at the end of a replication
  - finalized at the end of a generation

In the future, it would be nice to implement that as an Observer
pattern to decouple the involved classes