Move post-treatment shared source files to new folder
This commit creates a new folder `libaevol_post` in the post-treatment
source folder, that is meant to host all the files that are shared
between several post-treatments.

The commit also adds a new CMakeLists.txt file in that folder, which
builds a library called `libaevol-post`, to be then linked to the
post-treatments that need it.

The first files moved there are the IndivAnalysis.{cpp, h} files.

Finally, the CMakeLists.txt file in the post-treatment folder is updated
to reflect those changes.