[Stats] Fix disappearance of stats first line
When restarting aevol_run in a folder, the old statistics files are
copied before aevol appends the new statistics.
In this process, the first line of data (but not the header!)
disappeared, meaning that:
  - when running aevol_run for the 1st time, data for generation 0
    (written by aevol_create) disappeared.
  - when restarting aevol_run for the nth time, data for generation n-1
    (and all earlier generations) disappeared.

While not very important as the first generations are not so meaningful,
this bug was quite puzzling.