[aevol_post_lineage] Fix bug with indiv_by_id
When running the `aevol_post_lineage` post-treatment, individuals
generated by ReplicationReports are periodically checked against
individuals recovered from backups to make sure everything is in order.

Up until now, these backed-up individuals were loaded with the
ExpManager::indiv_by_id() method. However, as clones can share the same
id, this method no longer works and should not be called anymore (and is
to be removed later on).

This commit addresses this by correctly loading the individual with the
World::indiv_at() method, which takes the x and y coordinates instead of
the id.